HBX-1998: Rename class 'org.hibernate.tool.internal.reveng.PrimaryKeyInfo' to 'org.hibernate.tool.internal.reveng.binder.PrimaryKeyInfo'
